Safer, more efficient way to get ES Index sizes

+++ src/main/java/uk/co/recipes/service/api/IPersistenceAPI.java
@@ -27,7 +27,7 @@ public interface IPersistenceAPI<T> {
 
     String toStringId( final T obj) throws IOException;
 
-    int countAll() throws IOException;
+    long countAll() throws IOException;
 
     void deleteAll() throws IOException;
 }
